+++ assets/utils/offering.ts
@@ -1,4 +1,64 @@
-import { OfferingTypeEnum } from "@/@types/app";
+import { OfferingTypeEnum, StoredDataTypeEnum, type Service } from "@/@types/app";
+import { routes } from "@/router/router";
+
+export const editableOfferingTypes: OfferingTypeEnum[] = [
+    OfferingTypeEnum.WFS,
+    OfferingTypeEnum.WMSVECTOR,
+    OfferingTypeEnum.WMSRASTER,
+    OfferingTypeEnum.WMTSTMS,
+];
+
+export function getServiceEditLink(datastoreId: string, datasheetName: string, service: Service) {
+    switch (service.type) {
+        case OfferingTypeEnum.WMSVECTOR:
+            return routes.datastore_wms_vector_service_edit({
+                datastoreId,
+                vectorDbId: service.configuration.type_infos.used_data[0].stored_data,
+                offeringId: service._id,
+                datasheetName,
+            }).link;
+
+        case OfferingTypeEnum.WMSRASTER:
+            return routes.datastore_pyramid_raster_wms_raster_service_edit({
+                datastoreId,
+                pyramidId: service.configuration.type_infos.used_data[0].stored_data,
+                offeringId: service._id,
+                datasheetName,
+            }).link;
+
+        case OfferingTypeEnum.WFS:
+            return routes.datastore_wfs_service_edit({
+                datastoreId,
+                vectorDbId: service.configuration.type_infos.used_data[0].stored_data,
+                offeringId: service._id,
+                datasheetName,
+            }).link;
+
+        case OfferingTypeEnum.WMTSTMS:
+            switch (service.configuration.pyramid?.type) {
+                case StoredDataTypeEnum.ROK4PYRAMIDVECTOR:
+                    return routes.datastore_pyramid_vector_tms_service_edit({
+                        datastoreId,
+                        pyramidId: service.configuration.type_infos.used_data[0].stored_data,
+                        offeringId: service._id,
+                        datasheetName,
+                    }).link;
+                case StoredDataTypeEnum.ROK4PYRAMIDRASTER:
+                    return routes.datastore_pyramid_raster_wmts_service_edit({
+                        datastoreId,
+                        pyramidId: service.configuration.type_infos.used_data[0].stored_data,
+                        offeringId: service._id,
+                        datasheetName,
+                    }).link;
+
+                default:
+                    return routes.page_not_found().link;
+            }
+
+        default:
+            return routes.page_not_found().link;
+    }
+}
 
 export const offeringTypeDisplayName = (type: OfferingTypeEnum): string => {
     switch (type) {
